Update history page
Move formatHistory and totalTime processing to server side
